+++ Libraries/LibMedia/CMakeLists.txt
@@ -10,6 +10,7 @@ set(SOURCES
     PlaybackStates/PausedStateHandler.cpp
     PlaybackStates/PlaybackStateHandler.cpp
     PlaybackStates/ResumingStateHandler.cpp
+    Processors/AudioMixer.cpp
     Providers/AudioDataProvider.cpp
     Providers/GenericTimeProvider.cpp
     Providers/VideoDataProvider.cpp
