+++ Libraries/LibMedia/PlaybackStates/SuspendedStateHandler.h
@@ -9,6 +9,7 @@
 #include <LibMedia/PlaybackManager.h>
 #include <LibMedia/PlaybackStates/Forward.h>
 #include <LibMedia/PlaybackStates/SeekingStateHandler.h>
+#include <LibMedia/Processors/AudioMixer.h>
 #include <LibMedia/Providers/AudioDataProvider.h>
 #include <LibMedia/Providers/VideoDataProvider.h>
 
@@ -85,7 +86,7 @@ public:
 
         VERIFY(track.type() == TrackType::Audio);
         auto& track_data = manager().get_audio_data_for_track(track);
-        if (!manager().m_audio_sink)
+        if (!manager().m_audio_mixer)
             return;
         track_data.provider->resume();
         track_data.provider->seek(manager().current_time(), [manager = manager().weak(), provider = track_data.provider] {
