+++ Libraries/LibMedia/Sinks/AudioPlaybackSink.cpp
@@ -7,40 +7,31 @@
 #include <AK/Array.h>
 #include <AK/Atomic.h>
 #include <AK/AtomicRefCounted.h>
-#include <AK/HashMap.h>
 #include <AK/Time.h>
 #include <LibMedia/Audio/PlaybackStream.h>
 #include <LibMedia/AudioBlock.h>
-#include <LibMedia/Providers/AudioDataProvider.h>
+#include <LibMedia/Processors/AudioMixer.h>
 #include <LibSync/ConditionVariable.h>
+#include <LibSync/Mutex.h>
 #include <LibThreading/Thread.h>
 
 #include "AudioPlaybackSink.h"
 
 namespace Media {
 
-static constexpr size_t MAX_SAMPLES_PER_OUTPUT_BLOCK = 1024;
 static constexpr size_t OUTPUT_BLOCK_QUEUE_CAPACITY = 4;
 
 class AudioPlaybackSink::OutputThreadData : public AtomicRefCounted<OutputThreadData> {
 public:
-    struct TrackMixingData {
-        TrackMixingData(NonnullRefPtr<AudioDataProvider> const& provider)
-            : provider(provider)
-        {
-        }
+    OutputThreadData(NonnullRefPtr<AudioMixer>&& mixer)
+        : m_mixer(move(mixer))
+    {
+    }
 
-        NonnullRefPtr<AudioDataProvider> provider;
-        AudioBlock current_block;
-        bool buffering { false };
-    };
-
-    OutputThreadData() = default;
-
-    bool mix_one_block_into(AudioBlock&);
     ReadonlySpan<float> move_output_to_playback_stream_buffer(Span<float>);
 
     RefPtr<Audio::PlaybackStream> m_playback_stream;
+    NonnullRefPtr<AudioMixer> m_mixer;
 
     mutable Sync::Mutex m_output_mutex;
     mutable Sync::ConditionVariable m_output_condition { m_output_mutex };
@@ -54,31 +45,14 @@ public:
     Atomic<bool> m_pause_writing_audio_data { true };
     bool m_filler_is_waiting_in_output_loop { false };
     bool m_filler_should_exit { false };
-
-    mutable Sync::Mutex m_mixing_data_mutex;
-    Audio::SampleSpecification m_sample_specification;
-    HashMap<Track, TrackMixingData> m_track_mixing_datas;
-    Atomic<i64, MemoryOrder::memory_order_relaxed> m_next_sample_to_write { 0 };
-
-    Function<void(Track const&)> on_track_started_buffering;
 };
 
-ErrorOr<NonnullRefPtr<AudioPlaybackSink>> AudioPlaybackSink::try_create()
+ErrorOr<NonnullRefPtr<AudioPlaybackSink>> AudioPlaybackSink::try_create(NonnullRefPtr<AudioMixer> mixer)
 {
-    auto weak_ref = TRY(try_make_ref_counted<AudioMixingSinkWeakReference>());
-    auto output_thread_data = TRY(adopt_nonnull_ref_or_enomem(new (nothrow) OutputThreadData));
-    auto sink = TRY(try_make_ref_counted<AudioPlaybackSink>(weak_ref, output_thread_data));
-    weak_ref->emplace(sink);
+    auto output_thread_data = TRY(adopt_nonnull_ref_or_enomem(new (nothrow) OutputThreadData(move(mixer))));
+    auto sink = TRY(try_make_ref_counted<AudioPlaybackSink>(output_thread_data));
 
-    output_thread_data->on_track_started_buffering = [weak_self = sink->m_weak_self, &event_loop = sink->m_main_thread_event_loop](Track const& track) {
-        event_loop.deferred_invoke([weak_self, track] {
-            auto self = weak_self->take_strong();
-            if (self && self->on_start_buffering)
-                self->on_start_buffering(track);
-        });
-    };
-
-    auto thread = TRY(Threading::Thread::try_create("Audio Mixer"sv,
+    auto thread = TRY(Threading::Thread::try_create("Audio Processor"sv,
         [output_thread_data]() -> intptr_t {
             while (true) {
                 size_t tail_index;
@@ -107,7 +81,7 @@ ErrorOr<NonnullRefPtr<AudioPlaybackSink>> AudioPlaybackSink::try_create()
                     tail_index = output_thread_data->m_block_tail;
                 }
 
-                if (!output_thread_data->mix_one_block_into(output_thread_data->m_blocks[tail_index]))
+                if (!output_thread_data->m_mixer->mix_one_block_into(output_thread_data->m_blocks[tail_index]))
                     continue;
 
                 {
@@ -128,54 +102,20 @@ ErrorOr<NonnullRefPtr<AudioPlaybackSink>> AudioPlaybackSink::try_create()
     return sink;
 }
 
-AudioPlaybackSink::AudioPlaybackSink(AudioMixingSinkWeakReference& weak_ref, NonnullRefPtr<OutputThreadData> output_thread_data)
+AudioPlaybackSink::AudioPlaybackSink(NonnullRefPtr<OutputThreadData> output_thread_data)
     : m_main_thread_event_loop(Core::EventLoop::current())
-    , m_weak_self(weak_ref)
     , m_output_thread_data(move(output_thread_data))
 {
-    m_main_thread_event_loop.deferred_invoke([weak_self = m_weak_self] {
-        auto self = weak_self->take_strong();
-        if (!self)
-            return;
+    m_main_thread_event_loop.deferred_invoke([self = NonnullRefPtr(*this)] {
         self->create_playback_stream();
     });
 }
 
 AudioPlaybackSink::~AudioPlaybackSink()
 {
-    {
-        Sync::MutexLocker locker { m_output_thread_data->m_output_mutex };
-        m_output_thread_data->m_filler_should_exit = true;
-        m_output_thread_data->m_output_condition.broadcast();
-    }
-    m_weak_self->revoke();
-}
-
-void AudioPlaybackSink::set_provider(Track const& track, RefPtr<AudioDataProvider> const& provider)
-{
-    {
-        Sync::MutexLocker locker { m_output_thread_data->m_mixing_data_mutex };
-        m_output_thread_data->m_track_mixing_datas.remove(track);
-        if (provider == nullptr)
-            return;
-
-        // The provider must have its output sample specification set before it starts decoding, or
-        // we'll drop some samples due to a mismatch.
-        m_output_thread_data->m_track_mixing_datas.set(track, OutputThreadData::TrackMixingData(*provider));
-    }
-
-    if (m_output_thread_data->m_sample_specification.is_valid()) {
-        provider->set_output_sample_specification(m_output_thread_data->m_sample_specification);
-        provider->start();
-    }
-}
-
-RefPtr<AudioDataProvider> AudioPlaybackSink::provider(Track const& track) const
-{
-    auto mixing_data = m_output_thread_data->m_track_mixing_datas.get(track);
-    if (!mixing_data.has_value())
-        return nullptr;
-    return mixing_data->provider;
+    Sync::MutexLocker locker { m_output_thread_data->m_output_mutex };
+    m_output_thread_data->m_filler_should_exit = true;
+    m_output_thread_data->m_output_condition.broadcast();
 }
 
 void AudioPlaybackSink::create_playback_stream()
@@ -192,19 +132,10 @@ void AudioPlaybackSink::create_playback_stream()
 
     auto promise = Audio::PlaybackStream::create(Audio::OutputState::Suspended, target_latency_ms, move(data_callback));
 
-    promise->when_resolved([weak_self = m_weak_self](auto& stream) {
-        auto self = weak_self->take_strong();
-        if (!self)
-            return;
-
+    promise->when_resolved([self = NonnullRefPtr(*this)](auto& stream) {
         self->m_output_thread_data->m_playback_stream = stream;
-        self->m_output_thread_data->m_sample_specification = stream->sample_specification();
+        self->m_output_thread_data->m_mixer->set_sample_specification(stream->sample_specification());
         self->set_volume(self->m_volume);
-
-        for (auto& [track, track_data] : self->m_output_thread_data->m_track_mixing_datas) {
-            track_data.provider->set_output_sample_specification(self->m_output_thread_data->m_sample_specification);
-            track_data.provider->start();
-        }
 
         if (self->m_temporary_time.has_value()) {
             self->set_time(self->m_temporary_time.release_value());
@@ -221,164 +152,10 @@ void AudioPlaybackSink::create_playback_stream()
             self->resume();
     });
 
-    promise->when_rejected([weak_self = m_weak_self](auto& error) {
-        auto self = weak_self->take_strong();
-        if (!self)
-            return;
-
+    promise->when_rejected([self = NonnullRefPtr(*this)](auto& error) {
         if (self->on_audio_output_error)
             self->on_audio_output_error(move(error));
     });
-}
-
-bool AudioPlaybackSink::OutputThreadData::mix_one_block_into(AudioBlock& out_block)
-{
-    VERIFY(m_sample_specification.is_valid());
-
-    auto channel_count = m_sample_specification.channel_count();
-    auto max_sample_count = MAX_SAMPLES_PER_OUTPUT_BLOCK / channel_count;
-
-    Sync::MutexLocker mixing_data_locker { m_mixing_data_mutex };
-    auto buffer_start = m_next_sample_to_write.load();
-    auto initial_samples_end = buffer_start + static_cast<i64>(max_sample_count);
-    auto samples_end = initial_samples_end;
-
-    auto buffering = false;
-    auto any_track_has_fresh_data = false;
-    for (auto& [track, track_data] : m_track_mixing_datas) {
-        auto available_end = track_data.provider->queue_end_sample();
-        // A newly-enabled track has no data at the current mix position yet; skip it for clamping so
-        // the mixer doesn't stall waiting for it to catch up.
-        if (available_end <= buffer_start) {
-            track_data.current_block.clear();
-            while (true) {
-                auto block = track_data.provider->retrieve_block();
-                if (block.is_empty())
-                    break;
-                if (block.end_timestamp_in_samples() >= buffer_start) {
-                    available_end = block.end_timestamp_in_samples();
-                    track_data.current_block = move(block);
-                    break;
-                }
-            }
-            if (track_data.current_block.is_empty())
-                continue;
-        }
-        any_track_has_fresh_data = true;
-        if (available_end < samples_end) {
-            samples_end = available_end;
-            if (track_data.provider->is_blocked())
-                buffering = true;
-        }
-    }
-
-    if (!m_track_mixing_datas.is_empty() && !any_track_has_fresh_data)
-        return false;
-
-    for (auto& [track, track_data] : m_track_mixing_datas) {
-        if (!buffering) {
-            track_data.buffering = false;
-        } else {
-            if (!track_data.provider->is_blocked())
-                continue;
-            if (track_data.buffering)
-                continue;
-            track_data.buffering = true;
-
-            if (on_track_started_buffering)
-                on_track_started_buffering(track);
-        }
-    }
-
-    auto sample_count = static_cast<size_t>(max(samples_end - buffer_start, 0));
-    auto write_size = sample_count * channel_count;
-
-    if (sample_count == 0)
-        return false;
-
-    out_block.emplace(m_sample_specification, buffer_start, [&](AudioBlock::Data& data) {
-        if (data.size() != write_size)
-            data = MUST(AudioBlock::Data::create(write_size));
-        for (size_t i = 0; i < write_size; i++)
-            data[i] = 0.0f;
-
-        for (auto& [track, track_data] : m_track_mixing_datas) {
-            auto next_sample = buffer_start;
-
-            auto go_to_next_block = [&] {
-                auto new_block = track_data.provider->retrieve_block();
-                if (new_block.is_empty())
-                    return false;
-
-                track_data.current_block = move(new_block);
-                return true;
-            };
-
-            if (track_data.current_block.is_empty()) {
-                if (!go_to_next_block())
-                    continue;
-            }
-
-            while (!track_data.current_block.is_empty()) {
-                auto& current_block = track_data.current_block;
-                auto current_block_sample_count = static_cast<i64>(current_block.sample_count());
-
-                if (current_block.sample_specification() != m_sample_specification) {
-                    if (!go_to_next_block())
-                        break;
-                    current_block.clear();
-                    continue;
-                }
-
-                auto first_sample_offset = current_block.timestamp_in_samples();
-                if (first_sample_offset >= samples_end)
-                    break;
-
-                auto block_end = first_sample_offset + current_block_sample_count;
-                if (block_end <= next_sample) {
-                    if (!go_to_next_block())
-                        break;
-                    continue;
-                }
-
-                next_sample = max(next_sample, first_sample_offset);
-
-                VERIFY(next_sample >= first_sample_offset);
-                auto index_in_block = static_cast<size_t>((next_sample - first_sample_offset) * channel_count);
-                VERIFY(index_in_block < current_block.data_count());
-
-                VERIFY(next_sample >= buffer_start);
-                auto index_in_buffer = static_cast<size_t>((next_sample - buffer_start) * channel_count);
-                VERIFY(index_in_buffer < write_size);
-
-                VERIFY(current_block.data_count() >= index_in_block);
-                auto write_count = current_block.data_count() - index_in_block;
-                write_count = min(write_count, write_size - index_in_buffer);
-                VERIFY(write_count > 0);
-                VERIFY(index_in_buffer + write_count <= write_size);
-                VERIFY(write_count % channel_count == 0);
-
-                for (size_t i = 0; i < write_count; i++)
-                    data[index_in_buffer + i] += current_block.data()[index_in_block + i];
-
-                auto write_end = index_in_block + write_count;
-                if (write_end == current_block.data_count()) {
-                    if (!go_to_next_block())
-                        break;
-                    continue;
-                }
-                VERIFY(write_end < current_block.data_count());
-
-                next_sample += static_cast<i64>(write_count / channel_count);
-                if (next_sample == samples_end)
-                    break;
-                VERIFY(next_sample < samples_end);
-            }
-        }
-    });
-
-    m_next_sample_to_write += static_cast<i64>(sample_count);
-    return true;
 }
 
 ReadonlySpan<float> AudioPlaybackSink::OutputThreadData::move_output_to_playback_stream_buffer(Span<float> buffer)
@@ -455,11 +232,7 @@ void AudioPlaybackSink::resume()
     if (!m_output_thread_data->m_playback_stream)
         return;
     m_output_thread_data->m_playback_stream->resume()
-        ->when_resolved([weak_self = m_weak_self](auto new_stream_time) {
-            auto self = weak_self->take_strong();
-            if (!self)
-                return;
-
+        ->when_resolved([self = NonnullRefPtr(*this)](auto new_stream_time) {
             self->m_main_thread_event_loop.deferred_invoke([self, new_stream_time]() {
                 auto new_media_time = self->m_last_media_time + (new_stream_time - self->m_last_stream_time);
                 self->m_last_stream_time = new_stream_time;
@@ -503,18 +276,14 @@ void AudioPlaybackSink::set_time(AK::Duration time)
     m_output_thread_data->m_pause_writing_audio_data = true;
 
     m_output_thread_data->m_playback_stream->drain_buffer_and_suspend()
-        ->when_resolved([weak_self = m_weak_self]() {
-            auto self = weak_self->take_strong();
-            if (!self)
-                return;
-
+        ->when_resolved([self = NonnullRefPtr(*this)]() {
             auto new_stream_time = self->m_output_thread_data->m_playback_stream->total_time_played();
 
             self->m_main_thread_event_loop.deferred_invoke([self, new_stream_time]() {
                 self->m_last_stream_time = new_stream_time;
                 self->m_last_media_time = self->m_temporary_time.release_value();
 
-                auto seek_target_in_samples = self->m_last_media_time.to_time_units(1, self->m_output_thread_data->m_sample_specification.sample_rate());
+                auto seek_target_in_samples = self->m_last_media_time.to_time_units(1, self->m_output_thread_data->m_mixer->sample_specification().sample_rate());
 
                 {
                     Sync::MutexLocker output_locker { self->m_output_thread_data->m_output_mutex };
@@ -524,9 +293,7 @@ void AudioPlaybackSink::set_time(AK::Duration time)
                     self->m_output_thread_data->m_block_tail = 0;
                     self->m_output_thread_data->m_block_count = 0;
 
-                    self->m_output_thread_data->m_next_sample_to_write = seek_target_in_samples;
-                    for (auto& [track, track_data] : self->m_output_thread_data->m_track_mixing_datas)
-                        track_data.current_block.clear();
+                    self->m_output_thread_data->m_mixer->reset_to_sample_position(seek_target_in_samples);
 
                     self->m_output_thread_data->m_next_sample_to_play = seek_target_in_samples;
