+++ Libraries/LibMedia/Sinks/AudioPlaybackSink.h
@@ -13,25 +13,17 @@
 #include <LibMedia/Export.h>
 #include <LibMedia/Forward.h>
 #include <LibMedia/Providers/MediaTimeProvider.h>
-#include <LibMedia/Sinks/AudioSink.h>
-#include <LibSync/Mutex.h>
 
 namespace Media {
 
-class MEDIA_API AudioPlaybackSink final : public AudioSink
-    , public MediaTimeProvider {
-    class AudioMixingSinkWeakReference;
-
+class MEDIA_API AudioPlaybackSink final : public MediaTimeProvider {
 private:
     class OutputThreadData;
 
 public:
-    static ErrorOr<NonnullRefPtr<AudioPlaybackSink>> try_create();
-    AudioPlaybackSink(AudioMixingSinkWeakReference&, NonnullRefPtr<OutputThreadData>);
+    static ErrorOr<NonnullRefPtr<AudioPlaybackSink>> try_create(NonnullRefPtr<AudioMixer>);
+    AudioPlaybackSink(NonnullRefPtr<OutputThreadData>);
     virtual ~AudioPlaybackSink() override;
-
-    virtual void set_provider(Track const&, RefPtr<AudioDataProvider> const&) override;
-    virtual RefPtr<AudioDataProvider> provider(Track const&) const override;
 
     virtual AK::Duration current_time() const override;
     virtual void resume() override;
@@ -41,32 +33,11 @@ public:
     void set_volume(double);
 
     Function<void(Error&&)> on_audio_output_error;
-    Function<void(Track const&)> on_start_buffering;
 
 private:
-    class AudioMixingSinkWeakReference : public AtomicRefCounted<AudioMixingSinkWeakReference> {
-    public:
-        void emplace(AudioPlaybackSink& sink) { m_ptr = &sink; }
-        RefPtr<AudioPlaybackSink> take_strong() const
-        {
-            Sync::MutexLocker locker { m_mutex };
-            return m_ptr;
-        }
-        void revoke()
-        {
-            Sync::MutexLocker locker { m_mutex };
-            m_ptr = nullptr;
-        }
-
-    private:
-        mutable Sync::Mutex m_mutex;
-        AudioPlaybackSink* m_ptr { nullptr };
-    };
-
     void create_playback_stream();
 
     Core::EventLoop& m_main_thread_event_loop;
-    NonnullRefPtr<AudioMixingSinkWeakReference> m_weak_self;
 
     bool m_started_creating_playback_stream { false };
     bool m_playing { false };
